// src/app/core/resources/dto/login-info.ts
/**
 * BWZ-Rappi - Bank of Rapperswil Backend
 * Das Backend für die Bank of Rapperswil stellt eine vereinfachte API für das Management von Finanz-Transaktionen als Übungsplattform bereit.
 *
 * OpenAPI spec version: v0.51.1
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Spezifiziert den Benutzernamen und das Password des Users.
 */
export interface LoginInfo { 
    /**
     * Spezifiziert den eindeutigen Login-Namen des Benutzers.
     */
    readonly login: string;
    /**
     * Enthält das vom Benutzer gewählte Passwort.
     */
    readonly password: string;
}