// src/app/core/resources/dto/registration-info.ts
/**
 * BWZ-Rappi - Bank of Rapperswil Backend
 * Das Backend für die Bank of Rapperswil stellt eine vereinfachte API für das Management von Finanz-Transaktionen als Übungsplattform bereit.
 *
 * OpenAPI spec version: v0.51.1
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Enthält die Registierungsinformationen zu einem neuen Account.
 */
export interface RegistrationInfo { 
    /**
     * Spezifiziert den eindeutigen Login-Namen des Benutzers.
     */
    readonly login: string;
    /**
     * Definiert den Vornamen für den neuen Account.
     */
    readonly firstname: string;
    /**
     * Definiert den Nachnamen für den neuen Account.
     */
    readonly lastname: string;
    /**
     * Enthält das vom Benutzer gewählte Passwort.
     */
    readonly password: string;
}